some script renaming for up-to-date eval script
