+++ scripts/eval.py
@@ -4,23 +4,11 @@ import lightning as L
 import json
 import omegaconf
 import torch
-import torch.utils._pytree as pytree
 import wandb
 
-from functools import partial
-from pathlib import Path
-import os
-
-import plotly.express as px
-import plotly.graph_objects as go
-from plotly.subplots import make_subplots
-
-from non_rigid.datasets.proc_cloth_flow import ProcClothFlowDataModule
-from non_rigid.utils.vis_utils import FlowNetAnimation
 from non_rigid.utils.script_utils import (
     create_model,
     create_datamodule,
-    flatten_outputs,
     load_checkpoint_config_from_wandb,
 )
 
@@ -29,10 +17,7 @@ from non_rigid.utils.pointcloud_utils import expand_pcd
 from tqdm import tqdm
 import numpy as np
 
-from pytorch3d.transforms import Transform3d
 import rpad.visualize_3d.plots as vpl
-
-
 
 def visualize_batched_point_clouds(point_clouds):
     """
@@ -50,9 +35,6 @@ def visualize_batched_point_clouds(point_clouds):
         torch.cat(pcs),
         torch.cat(segs),
     )
-
-
-
 
 @torch.no_grad()
 @hydra.main(config_path="../configs", config_name="eval", version_base="1.3")
@@ -89,12 +71,24 @@ def main(cfg):
     device = f"cuda:{cfg.resources.gpus[0]}"
 
     ######################################################################
-    # Create the datamodule.
-    # Should be the same one as in training, but we're gonna use val+test
-    # dataloaders.
+    # Manually setting eval-specific configs.
     ######################################################################
-    # TODO: this is only used to compute the coverage metric, so this could be moved
-    # into the coverage block to avoid creating unnecessary datamodules
+    # Using a custom cloth-specific batch size, to allow for simultaneous evaluation 
+    # of RMSE, coverage, and precision.
+    if cfg.dataset.hole == "single":
+        bs = 1
+    elif cfg.dataset.hole == "double":
+        bs = 2
+    else:
+        raise ValueError(f"Unknown hole type: {cfg.dataset.hole}.")
+    cfg.inference.batch_size = bs
+    cfg.inference.val_batch_size = bs
+
+    ######################################################################
+    # Create the datamodule. This is just to initialize the datasets - we are
+    # not going to use the dataloaders, because we need to manually downsample 
+    # and batch.
+    ######################################################################
     cfg, datamodule = create_datamodule(cfg)
 
     ######################################################################
@@ -108,7 +102,6 @@ def main(cfg):
     # Model architecture is dataset-dependent, so we have a helper
     # function to create the model (while separating out relevant vals).
     network, model = create_model(cfg)
-
 
     # get checkpoint file (for now, this does not log a run)
     checkpoint_reference = cfg.checkpoint.reference
@@ -128,163 +121,82 @@ def main(cfg):
     network.eval()
     model.eval()
 
-
     ######################################################################
-    # Create the trainer.
-    # Bit of a misnomer here, we're not doing training. But we are gonna
-    # use it to set up the model appropriately and do all the batching
-    # etc.
-    #
-    # If this is a different kind of downstream eval, chuck this block.
+    # Helper function to run evals for a given dataset.
     ######################################################################
+    def run_eval(dataset, model):
+        num_samples = cfg.inference.num_wta_trials // bs
+        num_batches = len(dataset) // bs
+        eval_keys = ["pc_action", "pc_anchor", "pc", "flow", "seg", "seg_anchor", "T_action2world", "T_goal2world"]
+        if cfg.model.rel_pose:
+            eval_keys.append("rel_pose")
+            
+        rmse = []
+        coverage = []
+        precision = []
 
-    trainer = L.Trainer(
-        accelerator="gpu",
-        devices=cfg.resources.gpus,
-        # precision="16-mixed",
-        precision="32-true",
-        logger=False,
-    )
+
+        for i in tqdm(range(num_batches)):
+            batch_list = []
+
+            # get first item in batch, and keep downsampling indices
+            item = dataset.__getitem__(i * bs, return_indices=True)
+            downsample_indices = {
+                "action_pc_indices": item["action_pc_indices"],
+                "anchor_pc_indices": item["anchor_pc_indices"],
+            }
+            batch_list.append({key: item[key] for key in eval_keys})
+
+            # get the rest of the batch
+            for j in range(1, bs):
+                item = dataset.__getitem__(i * bs + j, downsample_indices=downsample_indices)
+                batch_list.append({key: item[key] for key in eval_keys})
+
+            # convert to batch
+            batch = {key: torch.stack([item[key] for item in batch_list]) for key in eval_keys}
+
+            # generate predictions
+            pred_dict = model.predict(batch, num_samples, progress=False)
+            pred_pc = pred_dict["point"]["pred"]
+            # pc = batch["pc"].to(device)
+            # seg = batch["seg"].to(device)
+
+            batch_rmse = torch.zeros(bs, cfg.inference.num_wta_trials)
+
+            for j in range(bs):
+                # expand ground truth pc to compute RMSE for cloth-specific sample
+                gt_pc = batch["pc"][j].unsqueeze(0).to(device)
+                seg = batch["seg"][j].unsqueeze(0).to(device)
+                gt_pc = expand_pcd(gt_pc, num_samples)
+                seg = expand_pcd(seg, num_samples)
+                batch_rmse[j] = flow_rmse(pred_pc, gt_pc, mask=True, seg=seg)
+
+            # computing precision and coverage
+            batch_precision = torch.min(batch_rmse, dim=0).values
+            batch_coverage = torch.min(batch_rmse, dim=1).values
+
+            # update dataset-wide metrics
+            rmse.append(batch_rmse.mean().item())
+            coverage.append(batch_coverage.mean().item())
+            precision.append(batch_precision.mean().item())
+            
+        rmse = np.mean(rmse)
+        coverage = np.mean(coverage)
+        precision = np.mean(precision)
+        return rmse, coverage, precision
+
 
     ######################################################################
     # Run the model on the train/val/test sets.
-    # This outputs a list of dictionaries, one for each batch. This
-    # is annoying to work with, so later we'll flatten.
-    #
-    # If a downstream eval, you can swap it out with whatever the eval
-    # function is.
     ######################################################################
-    
-    if cfg.coverage:
-        train_outputs, val_outputs, val_ood_outputs = trainer.predict(
-            model,
-            dataloaders=[
-                datamodule.train_dataloader(),
-                *datamodule.val_dataloader(),
-            ]
-            )
+    model.to(device)
+    train_rmse, train_coverage, train_precision = run_eval(datamodule.train_dataset, model)
+    val_rmse, val_coverage, val_precision = run_eval(datamodule.val_dataset, model)
+    val_ood_rmse, val_ood_coverage, val_ood_precision = run_eval(datamodule.val_ood_dataset, model)
 
-        for outputs_list, name in [
-            (train_outputs, "train"),
-            (val_outputs, "val"),
-            (val_ood_outputs, "val_ood")
-        ]:
-            # Put everything on CPU, and flatten a list of dicts into one dict.
-            out_cpu = [pytree.tree_map(lambda x: x.cpu(), o) for o in outputs_list]
-            outputs = flatten_outputs(out_cpu)
-            
-            # Compute the metrics.
-            rmse = torch.mean(outputs["rmse"])
-            rmse_wta = torch.mean(outputs["rmse_wta"])
-            print(f"{name} rmse: {rmse}, rmse wta: {rmse_wta}")
-
-
-    if cfg.precision:
-        model.to(device)
-        # precision_dm = 2
-
-        # TODO: this code chunk below currently breaks because all the datasets were 
-        # trained as "real", even though they were "dedo"...sad. can find a way to fix later
-        # # precision metric only for DEDO tasks
-        # if cfg.dataset.source != "dedo":
-        #     raise ValueError("Precision metric only implemented for DEDO tasks.")
-        
-        # determining number of samples and batch size based on cloth type
-        num_samples = 20
-        # NOTE: in theory, I would compare against all the cloths in the single set
-        # However, this kind of implementation would not extend well once I start 
-        # conditioning on anchor pose - in other words, using the whole train set is not analogous 
-        # to bucketizing the multi-cloth experiments, because the anchor pose is fixed in those 
-        # buckets - so it's better to bucketize here as well, and have higher precision rmses, 
-        # since we won't report them anyways
-        if cfg.dataset.hole == "single":
-            bs = 1
-        elif cfg.dataset.hole == "double":
-            bs = 2
-        else:
-            raise ValueError(f"Unknown hole type: {cfg.dataset.hole}.")
-        
-        # NOTE: this is very brittle - this will if a batch contains more than one cloth geometry
-        # consider setting this back after precision is computed so the whole script can be run together
-        
-        # extracting action sample size for manual downsampling
-        if bs > 1:
-            sample_size_action = cfg.dataset.sample_size_action
-            cfg.dataset.sample_size_action = -1
-
-        # otherwise, can downsample as usual
-        cfg.inference.batch_size = bs
-        cfg.inference.val_batch_size = bs
-
-
-
-        # set up precision datamodule
-        cfg, datamodule = create_datamodule(cfg)
-        # # set up a new datamodule for the dataloaders
-        # datamodule = ProcClothFlowDataModule(
-        #     batch_size=bs,
-        #     val_batch_size=bs,
-        #     num_workers=cfg.resources.num_workers,
-        #     dataset_cfg=cfg.dataset,
-        # )
-        # datamodule.setup(stage="predict")
-        train_loader = datamodule.train_dataloader()
-        val_loader, val_ood_loader = datamodule.val_dataloader()
-
-        # helper function to run precision evaluation
-        def precision_eval(dataloader, model):
-            precision_rmses = []
-            for batch in tqdm(dataloader):
-                # manually downsampling to ensure correspondences are consistent across the batch
-                # generate predictions
-                pred_dict = model.predict(batch, num_samples, progress=False)
-                pred_pc = pred_dict["point"]["pred"]
-                pc = batch["pc"].to(device)
-                seg = batch["seg"].to(device)
-
-                for pred in pred_pc:
-                    # expand each prediction to batch size, and save minimum RMSE
-                    pred = expand_pcd(pred.unsqueeze(0), bs)
-                    precision_rmses.append(torch.min(flow_rmse(pred, pc, mask=True, seg=seg)))
-                    
-                # VISUALIZATION CODE - KEEP FOR LATER
-                # pred_pc = pred_dict["point"]["pred_world"].cpu().numpy()
-                # # convert batch back to world frame
-                # T_action2world = Transform3d(
-                #     matrix=batch["T_action2world"]
-                # )
-                # T_goal2world = Transform3d(
-                #     matrix=batch["T_goal2world"]
-                # )
-                # pc_action_world = T_action2world.transform_points(batch["pc_action"]).numpy().reshape(-1, 3)
-                # pc_anchor_world = T_goal2world.transform_points(batch["pc_anchor"]).numpy().reshape(-1, 3)
-
-
-                # pc_action_seg = np.ones((pc_action_world.shape[0])) * 0
-                # pc_anchor_seg = np.ones((pc_anchor_world.shape[0])) * 1
-                # pred_pc_seg = np.ones((pred_pc.shape[0], pred_pc.shape[1])) * np.arange(pred_pc.shape[0]).reshape(-1, 1) + 2
-                
-                # vpl.segmentation_fig(
-                #     np.concatenate([
-                #         pc_action_world,
-                #         pc_anchor_world,
-                #         pred_pc.reshape(-1, 3),
-                #     ]),
-                #     np.concatenate([
-                #         pc_action_seg,
-                #         pc_anchor_seg,
-                #         pred_pc_seg.flatten(),
-                #     ]).astype(int),
-            
-                # ).show()
-            return torch.stack(precision_rmses).mean().item()
-
-        train_precision_rmse = precision_eval(train_loader, model)
-        val_precision_rmse = precision_eval(val_loader, model)
-        val_ood_precision_rmse = precision_eval(val_ood_loader, model)
-        print("Train Precision RMSE: ", train_precision_rmse)
-        print("Val Precision RMSE: ", val_precision_rmse)
-        print("Val OOD Precision RMSE: ", val_ood_precision_rmse)
+    print(f"Train RMSE: {train_rmse}, Coverage: {train_coverage}, Precision: {train_precision}")
+    print(f"Val RMSE: {val_rmse}, Coverage: {val_coverage}, Precision: {val_precision}")
+    print(f"Val OOD RMSE: {val_ood_rmse}, Coverage: {val_ood_coverage}, Precision: {val_ood_precision}")
 
 if __name__ == "__main__":
     main()
